+++ Remote/Data/MidiEventInfo.cs
@@ -1,6 +1,6 @@
 ﻿namespace MilkwaveRemote.Data {
   public class MidiEventInfo {
-    public int Value { get; set; } 
+    public int Value { get; set; }
     public int Channel { get; set; }
     public int Controller { get; set; }