+++ Remote/Data/TagEntry.cs
@@ -1,15 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MilkwaveRemote.Data {
+﻿namespace MilkwaveRemote.Data {
   public class TagEntry {
     public TagEntry() { }
 
     public String PresetPath { get; set; } = "";
-    
-    public List<String> Tags { get; set; } =  new List<String>();
+
+    public List<String> Tags { get; set; } = new List<String>();
   }
 }